feat : 실시간 인기 키워드 클릭시 최근 검색 기록에 남게하는 기능 구현
